// Calendar/Calendar/types/Keys.ts
export interface Keys {
    id: string;
    name: string;
    start: string;
    end: string;
    eventColor?: string;
    description?: string
    resource?: string;
    resourceName?: string;
    resourceId?: string;
    resourceEtn?: string;
    resourceEntitySetName?: string;
    resourceNavigationProperty?: string;
    resourceGetAllInModel?: boolean;
  }